# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineCreateRequest": {
        "additionalProperties": false,
        "description": "Body of ``POST /api/v1/pipelines``.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 1024,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "pipeline_id": {
            "maxLength": 68,
            "minLength": 7,
            "pattern": "^pip_[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Pipeline Id",
            "type": "string"
          }
        },
        "required": [
          "pipeline_id",
          "display_name"
        ],
        "title": "PipelineCreateRequest",
        "type": "object"
      },
      "PipelinePageResponse": {
        "additionalProperties": false,
        "description": "Paginated pipeline collection.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PipelineResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "next_cursor"
        ],
        "title": "PipelinePageResponse",
        "type": "object"
      },
      "PipelineResponse": {
        "additionalProperties": false,
        "description": "One pipeline identity with its mutable display metadata.",
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "pipeline_id": {
            "title": "Pipeline Id",
            "type": "string"
          },
          "row_version": {
            "title": "Row Version",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "pipeline_id",
          "display_name",
          "description",
          "created_at",
          "archived_at",
          "row_version"
        ],
        "title": "PipelineResponse",
        "type": "object"
      },
      "PipelineVersionAckResponse": {
        "additionalProperties": false,
        "description": "Publication acknowledgement without the specification body.\n\nThe published specification embeds connector binding snapshots whose\nfield names cannot pass the durable redaction wall applied to stored\nidempotent responses; the full document is addressable through the\nversion route.",
        "properties": {
          "pipeline_id": {
            "title": "Pipeline Id",
            "type": "string"
          },
          "planner_format_version": {
            "title": "Planner Format Version",
            "type": "integer"
          },
          "published_at": {
            "title": "Published At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "specification_sha256": {
            "title": "Specification Sha256",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "pipeline_id",
          "version",
          "specification_sha256",
          "planner_format_version",
          "published_at"
        ],
        "title": "PipelineVersionAckResponse",
        "type": "object"
      },
      "PipelineVersionFrontierResponse": {
        "additionalProperties": false,
        "description": "Immutable latest-version frontier for one pipeline.\n\n``latest_version`` is the newest immutable published version and is\n``0`` exactly when the pipeline has never been published. It is the\nvalue clients echo as ``expected_latest_version`` when publishing.\nMutable metadata row versions are deliberately not part of this\nresponse: they never fence publication.",
        "properties": {
          "latest_version": {
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Latest Version",
            "type": "integer"
          },
          "pipeline_id": {
            "title": "Pipeline Id",
            "type": "string"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "pipeline_id",
          "latest_version",
          "published"
        ],
        "title": "PipelineVersionFrontierResponse",
        "type": "object"
      },
      "PipelineVersionPublishRequest": {
        "additionalProperties": false,
        "description": "Body of ``POST /api/v1/pipelines/{id}/versions``.",
        "properties": {
          "document": {
            "additionalProperties": true,
            "title": "Document",
            "type": "object"
          },
          "expected_latest_version": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Latest Version"
          }
        },
        "required": [
          "document"
        ],
        "title": "PipelineVersionPublishRequest",
        "type": "object"
      },
      "PipelineVersionResponse": {
        "additionalProperties": false,
        "description": "One immutable published pipeline version.",
        "properties": {
          "pipeline_id": {
            "title": "Pipeline Id",
            "type": "string"
          },
          "planner_format_version": {
            "title": "Planner Format Version",
            "type": "integer"
          },
          "published_at": {
            "title": "Published At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "specification": {
            "additionalProperties": true,
            "title": "Specification",
            "type": "object"
          },
          "specification_sha256": {
            "title": "Specification Sha256",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "pipeline_id",
          "version",
          "specification",
          "specification_sha256",
          "planner_format_version",
          "published_at"
        ],
        "title": "PipelineVersionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ParityGrid",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pipelines": {
      "get": {
        "operationId": "list_pipelines_api_v1_pipelines_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelinePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pipelines",
        "tags": [
          "pipelines"
        ]
      },
      "post": {
        "operationId": "create_pipeline_api_v1_pipelines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pipeline",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/api/v1/pipelines/{pipeline_id}": {
      "get": {
        "operationId": "get_pipeline_api_v1_pipelines__pipeline_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/api/v1/pipelines/{pipeline_id}/version-frontier": {
      "get": {
        "operationId": "get_pipeline_version_frontier_api_v1_pipelines__pipeline_id__version_frontier_get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineVersionFrontierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Version Frontier",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/api/v1/pipelines/{pipeline_id}/versions": {
      "post": {
        "operationId": "publish_pipeline_version_api_v1_pipelines__pipeline_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineVersionPublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineVersionAckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Pipeline Version",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/api/v1/pipelines/{pipeline_id}/versions/{version}": {
      "get": {
        "operationId": "get_pipeline_version_api_v1_pipelines__pipeline_id__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Version",
        "tags": [
          "pipelines"
        ]
      }
    },
...
  },
...
}